Empty email address when I choose guest checkout
In case of guest checkout there is no customer entity in database, so properties like email address have to be filled in.
`_setCheckoutMethod()` copes the email address from the billing address, but it's empty, because it will be filled out later in the `_addAddressesToQuote()` method.
I swapped these two methods to get the email address filled in. 
But it might break something else. Please check it.